Moved file functions from SiteBuilderService class to new FileService class.
